+++ src/app/data/api/information.ts
@@ -1,15 +1,7 @@
 import { Information } from '../common/types';
+import { getValidateData } from './gateway';
 
 export async function getInformations() {
-  const response = await fetch('https://tyrael.up.railway.app/informations', {
-    cache: 'no-store',
-    method: 'GET',
-    headers: {
-      'Content-Type': 'application/json',
-      'Authorization': `Basic ${btoa('root:1234')}`,
-    },
-  });
-  
-  const data: Information[] = await (response.json());
+  const data: Information[] = await getValidateData('informations');
   return data;
 }