+++ src/app/data/api/skills.ts
@@ -1,31 +1,14 @@
 import { Skills } from '../common/types';
 import { getSkillIcon } from './common';
+import { getValidateData } from './gateway';
 
 export async function getSkills() {
-  const response = await fetch('https://tyrael.up.railway.app/skills', {
-    cache: 'no-store',
-    method: 'GET',
-    headers: {
-      'Content-Type': 'application/json',
-      'Authorization': `Basic ${btoa('root:1234')}`,
-    },
-  });
-  
-  const data: Skills[] = await (response.json());
+  const data: Skills[] = await getValidateData('skills');
   return data;
 }
 
 export async function getSkillsWithIcons() {
-  const response = await fetch('https://tyrael.up.railway.app/skills', {
-    cache: 'no-store',
-    method: 'GET',
-    headers: {
-      'Content-Type': 'application/json',
-      'Authorization': `Basic ${btoa('root:1234')}`,
-    },
-  });
-  
-  const data: Skills[] = await (response.json());
+  const data: Skills[] = await getValidateData('skills');
   for (let i = 0; i < data.length; i++) {
     const skills: Skills = data[i];
     skills.icon = getSkillIcon(skills.name);